修改gif制作帧率；修改视频画面提取频率

+++ app/src/main/java/com/suixue/edu/college/gifEncoder/GitEncoderUtil.java
@@ -34,7 +34,7 @@ public class GitEncoderUtil {
             throw new RuntimeException("input video file not found: " + videoPath);
         }
         desiredFrameRate = frameRate == 0 ? 10 : frameRate;
-        frameIntervalUs = 1000 * 1000 / desiredFrameRate;
+        frameIntervalUs = (int) (1000 * 1000f / desiredFrameRate);
         new Thread(new Runnable() {
             @Override
             public void run() {
